requirements.txt
flask==3.0.0
google-genai
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
requests==2.32.3
openai
